split user roles to its own component
